+++ ml-agents/mlagents/trainers/env_manager.py
@@ -8,7 +8,6 @@ from mlagents_envs.base_env import (
 )
 from mlagents_envs.side_channel.stats_side_channel import StatsAggregationMethod
 
-from mlagents.trainers.brain import BrainParameters
 from mlagents.trainers.policy.tf_policy import TFPolicy
 from mlagents.trainers.agent_processor import AgentManager, AgentManagerQueue
 from mlagents.trainers.action_info import ActionInfo
@@ -71,14 +70,14 @@ class EnvManager(ABC):
     def set_env_parameters(self, config: Dict = None) -> None:
         """
         Sends environment parameter settings to C# via the
-        EnvironmentParametersSidehannel.
+        EnvironmentParametersSideChannel.
         :param config: Dict of environment parameter keys and values
         """
         pass
 
     @property
     @abstractmethod
-    def external_brains(self) -> Dict[BehaviorName, BrainParameters]:
+    def training_behaviors(self) -> Dict[BehaviorName, BehaviorSpec]:
         pass
 
     @abstractmethod
@@ -93,7 +92,7 @@ class EnvManager(ABC):
             self._process_step_infos(self.first_step_infos)
             self.first_step_infos = None
         # Get new policies if found. Always get the latest policy.
-        for brain_name in self.external_brains:
+        for brain_name in self.training_behaviors:
             _policy = None
             try:
                 # We make sure to empty the policy queue before continuing to produce steps.