+++ ml-agents/mlagents/trainers/models.py
@@ -5,7 +5,6 @@ import numpy as np
 from mlagents.tf_utils import tf
 
 from mlagents.trainers.exception import UnityTrainerException
-from mlagents.trainers.brain import CameraResolution
 
 ActivationFunction = Callable[[tf.Tensor], tf.Tensor]
 EncoderFunction = Callable[
@@ -13,6 +12,12 @@ EncoderFunction = Callable[
 ]
 
 EPSILON = 1e-7
+
+
+class Tensor3DShape(NamedTuple):
+    height: int
+    width: int
+    num_channels: int
 
 
 class EncoderType(Enum):
@@ -90,9 +95,7 @@ class ModelUtils:
         return tf.multiply(input_activation, tf.nn.sigmoid(input_activation))
 
     @staticmethod
-    def create_visual_input(
-        camera_parameters: CameraResolution, name: str
-    ) -> tf.Tensor:
+    def create_visual_input(camera_parameters: Tensor3DShape, name: str) -> tf.Tensor:
         """
         Creates image input op.
         :param camera_parameters: Parameters for visual observation.
@@ -109,22 +112,40 @@ class ModelUtils:
         return visual_in
 
     @staticmethod
-    def create_visual_input_placeholders(
-        camera_resolutions: List[CameraResolution]
-    ) -> List[tf.Tensor]:
+    def create_input_placeholders(
+        observation_shapes: List[Tuple], name_prefix: str = ""
+    ) -> Tuple[tf.Tensor, List[tf.Tensor]]:
         """
         Creates input placeholders for visual inputs.
-        :param camera_resolutions: A List of CameraResolutions that specify the resolutions
-        of the input visual observations.
+        :param observation_shapes: A List of tuples that specify the resolutions
+            of the input observations. Tuples for now are restricted to 1D (vector) or 3D (Tensor)
+        :param name_prefix: A name prefix to add to the placeholder names. This is used so that there
+            is no conflict when creating multiple placeholder sets.
         :returns: A List of Tensorflow placeholders where the input iamges should be fed.
         """
         visual_in: List[tf.Tensor] = []
-        for i, camera_resolution in enumerate(camera_resolutions):
-            visual_input = ModelUtils.create_visual_input(
-                camera_resolution, name="visual_observation_" + str(i)
-            )
-            visual_in.append(visual_input)
-        return visual_in
+        vector_in_size = 0
+        for i, dimension in enumerate(observation_shapes):
+            if len(dimension) == 3:
+                _res = Tensor3DShape(
+                    height=dimension[0], width=dimension[1], num_channels=dimension[2]
+                )
+                visual_input = ModelUtils.create_visual_input(
+                    _res, name=name_prefix + "visual_observation_" + str(i)
+                )
+                visual_in.append(visual_input)
+            elif len(dimension) == 1:
+                vector_in_size += dimension[0]
+            else:
+                raise UnityTrainerException(
+                    f"Unsupported shape of {dimension} for observation {i}"
+                )
+        vector_in = tf.placeholder(
+            shape=[None, vector_in_size],
+            dtype=tf.float32,
+            name=name_prefix + "vector_observation",
+        )
+        return vector_in, visual_in
 
     @staticmethod
     def create_vector_input(
@@ -571,7 +592,8 @@ class ModelUtils:
                     )
                     visual_encoders.append(encoded_visual)
                 hidden_visual = tf.concat(visual_encoders, axis=1)
-            if vector_in.get_shape()[-1] > 0:  # Don't encode 0-shape inputs
+            if vector_in.get_shape()[-1] > 0:
+                # Don't encode non-existant or 0-shape inputs
                 hidden_state = ModelUtils.create_vector_observation_encoder(
                     vector_observation_input,
                     h_size,