+++ ml-agents/mlagents/trainers/tests/test_ppo.py
@@ -10,14 +10,11 @@ from mlagents.trainers.behavior_id_utils import BehaviorIdentifiers
 from mlagents.trainers.ppo.trainer import PPOTrainer, discount_rewards
 from mlagents.trainers.ppo.optimizer import PPOOptimizer
 from mlagents.trainers.policy.nn_policy import NNPolicy
-from mlagents.trainers.brain import BrainParameters
 from mlagents.trainers.agent_processor import AgentManagerQueue
 from mlagents.trainers.tests import mock_brain as mb
-from mlagents.trainers.tests.mock_brain import make_brain_parameters
 from mlagents.trainers.tests.test_trajectory import make_fake_trajectory
-from mlagents.trainers.settings import NetworkSettings, TrainerSettings, PPOSettings
+from mlagents.trainers.settings import NetworkSettings
 from mlagents.trainers.tests.test_simple_rl import PPO_CONFIG
-from mlagents.trainers.exception import TrainerConfigError
 from mlagents.trainers.tests.test_reward_signals import (  # noqa: F401; pylint: disable=unused-variable
     curiosity_dummy_config,
     gail_dummy_config,
@@ -29,7 +26,7 @@ def dummy_config():
     return copy.deepcopy(PPO_CONFIG)
 
 
-VECTOR_ACTION_SPACE = [2]
+VECTOR_ACTION_SPACE = 2
 VECTOR_OBS_SPACE = 8
 DISCRETE_ACTION_SPACE = [3, 3, 3, 2]
 BUFFER_INIT_SAMPLES = 64
@@ -37,12 +34,13 @@ NUM_AGENTS = 12
 
 
 def _create_ppo_optimizer_ops_mock(dummy_config, use_rnn, use_discrete, use_visual):
-    mock_brain = mb.setup_mock_brain(
+    mock_specs = mb.setup_test_behavior_specs(
         use_discrete,
         use_visual,
-        vector_action_space=VECTOR_ACTION_SPACE,
+        vector_action_space=DISCRETE_ACTION_SPACE
+        if use_discrete
+        else VECTOR_ACTION_SPACE,
         vector_obs_space=VECTOR_OBS_SPACE,
-        discrete_action_space=DISCRETE_ACTION_SPACE,
     )
 
     trainer_settings = attr.evolve(dummy_config)
@@ -52,33 +50,10 @@ def _create_ppo_optimizer_ops_mock(dummy_config, use_rnn, use_discrete, use_visu
         else None
     )
     policy = NNPolicy(
-        0, mock_brain, trainer_settings, False, "test", False, create_tf_graph=False
+        0, mock_specs, trainer_settings, False, "test", False, create_tf_graph=False
     )
     optimizer = PPOOptimizer(policy, trainer_settings)
     return optimizer
-
-
-def _create_fake_trajectory(use_discrete, use_visual, time_horizon):
-    if use_discrete:
-        act_space = DISCRETE_ACTION_SPACE
-    else:
-        act_space = VECTOR_ACTION_SPACE
-
-    if use_visual:
-        num_vis_obs = 1
-        vec_obs_size = 0
-    else:
-        num_vis_obs = 0
-        vec_obs_size = VECTOR_OBS_SPACE
-
-    trajectory = make_fake_trajectory(
-        length=time_horizon,
-        max_step_complete=True,
-        vec_obs_size=vec_obs_size,
-        num_vis_obs=num_vis_obs,
-        action_space=act_space,
-    )
-    return trajectory
 
 
 @pytest.mark.parametrize("discrete", [True, False], ids=["discrete", "continuous"])
@@ -91,7 +66,9 @@ def test_ppo_optimizer_update(dummy_config, rnn, visual, discrete):
         dummy_config, use_rnn=rnn, use_discrete=discrete, use_visual=visual
     )
     # Test update
-    update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, optimizer.policy.behavior_spec
+    )
     # Mock out reward signal eval
     update_buffer["advantages"] = update_buffer["environment_rewards"]
     update_buffer["extrinsic_returns"] = update_buffer["environment_rewards"]
@@ -116,7 +93,9 @@ def test_ppo_optimizer_update_curiosity(
         dummy_config, use_rnn=rnn, use_discrete=discrete, use_visual=visual
     )
     # Test update
-    update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, optimizer.policy.behavior_spec
+    )
     # Mock out reward signal eval
     update_buffer["advantages"] = update_buffer["environment_rewards"]
     update_buffer["extrinsic_returns"] = update_buffer["environment_rewards"]
@@ -138,7 +117,9 @@ def test_ppo_optimizer_update_gail(gail_dummy_config, dummy_config):  # noqa: F8
         PPO_CONFIG, use_rnn=False, use_discrete=False, use_visual=False
     )
     # Test update
-    update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, optimizer.policy.behavior_spec
+    )
     # Mock out reward signal eval
     update_buffer["advantages"] = update_buffer["environment_rewards"]
     update_buffer["extrinsic_returns"] = update_buffer["environment_rewards"]
@@ -151,7 +132,7 @@ def test_ppo_optimizer_update_gail(gail_dummy_config, dummy_config):  # noqa: F8
     )
 
     # Check if buffer size is too big
-    update_buffer = mb.simulate_rollout(3000, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(3000, optimizer.policy.behavior_spec)
     # Mock out reward signal eval
     update_buffer["advantages"] = update_buffer["environment_rewards"]
     update_buffer["extrinsic_returns"] = update_buffer["environment_rewards"]
@@ -174,7 +155,13 @@ def test_ppo_get_value_estimates(dummy_config, rnn, visual, discrete):
         dummy_config, use_rnn=rnn, use_discrete=discrete, use_visual=visual
     )
     time_horizon = 15
-    trajectory = _create_fake_trajectory(discrete, visual, time_horizon)
+    trajectory = make_fake_trajectory(
+        length=time_horizon,
+        observation_shapes=optimizer.policy.behavior_spec.observation_shapes,
+        max_step_complete=True,
+        action_space=DISCRETE_ACTION_SPACE if discrete else VECTOR_ACTION_SPACE,
+        is_discrete=discrete,
+    )
     run_out, final_value_out = optimizer.get_trajectory_value_estimates(
         trajectory.to_agentbuffer(), trajectory.next_obs, done=False
     )
@@ -215,28 +202,17 @@ def test_trainer_increment_step(ppo_optimizer):
     mock_optimizer.reward_signals = {}
     ppo_optimizer.return_value = mock_optimizer
 
-    brain_params = BrainParameters(
-        brain_name="test_brain",
-        vector_observation_space_size=1,
-        camera_resolutions=[],
-        vector_action_space_size=[2],
-        vector_action_descriptions=[],
-        vector_action_space_type=0,
-    )
-
-    trainer = PPOTrainer(
-        brain_params.brain_name, 0, trainer_params, True, False, 0, "0"
-    )
+    trainer = PPOTrainer("test_brain", 0, trainer_params, True, False, 0, "0")
     policy_mock = mock.Mock(spec=NNPolicy)
     policy_mock.get_current_step.return_value = 0
     step_count = (
         5  # 10 hacked because this function is no longer called through trainer
     )
     policy_mock.increment_step = mock.Mock(return_value=step_count)
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(brain_params.brain_name)
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
     trainer.add_policy(behavior_id, policy_mock)
 
-    trainer._increment_step(5, brain_params.brain_name)
+    trainer._increment_step(5, trainer.brain_name)
     policy_mock.increment_step.assert_called_with(5)
     assert trainer.step == step_count
 
@@ -245,12 +221,13 @@ def test_trainer_increment_step(ppo_optimizer):
 def test_trainer_update_policy(
     dummy_config, curiosity_dummy_config, use_discrete  # noqa: F811
 ):
-    mock_brain = mb.setup_mock_brain(
+    mock_behavior_spec = mb.setup_test_behavior_specs(
         use_discrete,
         False,
-        vector_action_space=VECTOR_ACTION_SPACE,
+        vector_action_space=DISCRETE_ACTION_SPACE
+        if use_discrete
+        else VECTOR_ACTION_SPACE,
         vector_obs_space=VECTOR_OBS_SPACE,
-        discrete_action_space=DISCRETE_ACTION_SPACE,
     )
 
     trainer_params = dummy_config
@@ -260,12 +237,13 @@ def test_trainer_update_policy(
 
     # Test curiosity reward signal
     trainer_params.reward_signals = curiosity_dummy_config
-    trainer = PPOTrainer(mock_brain.brain_name, 0, trainer_params, True, False, 0, "0")
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(mock_brain.brain_name)
-    policy = trainer.create_policy(behavior_id, mock_brain)
+    mock_brain_name = "MockBrain"
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(mock_brain_name)
+    trainer = PPOTrainer("test", 0, trainer_params, True, False, 0, "0")
+    policy = trainer.create_policy(behavior_id, mock_behavior_spec)
     trainer.add_policy(behavior_id, policy)
     # Test update with sequence length smaller than batch size
-    buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, mock_brain)
+    buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, mock_behavior_spec)
     # Mock out reward signal eval
     buffer["extrinsic_rewards"] = buffer["environment_rewards"]
     buffer["extrinsic_returns"] = buffer["environment_rewards"]
@@ -280,26 +258,24 @@ def test_trainer_update_policy(
 
 
 def test_process_trajectory(dummy_config):
-    brain_params = BrainParameters(
-        brain_name="test_brain",
-        vector_observation_space_size=1,
-        camera_resolutions=[],
-        vector_action_space_size=[2],
-        vector_action_descriptions=[],
-        vector_action_space_type=0,
+    behavior_spec = mb.setup_test_behavior_specs(
+        True,
+        False,
+        vector_action_space=DISCRETE_ACTION_SPACE,
+        vector_obs_space=VECTOR_OBS_SPACE,
     )
-    trainer = PPOTrainer(brain_params, 0, dummy_config, True, False, 0, "0")
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(brain_params.brain_name)
-    policy = trainer.create_policy(behavior_id, brain_params)
+    mock_brain_name = "MockBrain"
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(mock_brain_name)
+    trainer = PPOTrainer("test_brain", 0, dummy_config, True, False, 0, "0")
+    policy = trainer.create_policy(behavior_id, behavior_spec)
     trainer.add_policy(behavior_id, policy)
     trajectory_queue = AgentManagerQueue("testbrain")
     trainer.subscribe_trajectory_queue(trajectory_queue)
     time_horizon = 15
     trajectory = make_fake_trajectory(
         length=time_horizon,
+        observation_shapes=behavior_spec.observation_shapes,
         max_step_complete=True,
-        vec_obs_size=1,
-        num_vis_obs=0,
         action_space=[2],
     )
     trajectory_queue.put(trajectory)
@@ -323,8 +299,7 @@ def test_process_trajectory(dummy_config):
     trajectory = make_fake_trajectory(
         length=time_horizon + 1,
         max_step_complete=False,
-        vec_obs_size=1,
-        num_vis_obs=0,
+        observation_shapes=behavior_spec.observation_shapes,
         action_space=[2],
     )
     trajectory_queue.put(trajectory)
@@ -339,19 +314,17 @@ def test_process_trajectory(dummy_config):
 
 @mock.patch("mlagents.trainers.ppo.trainer.PPOOptimizer")
 def test_add_get_policy(ppo_optimizer, dummy_config):
-    brain_params = make_brain_parameters(
-        discrete_action=False, visual_inputs=0, vec_obs_size=6
-    )
     mock_optimizer = mock.Mock()
     mock_optimizer.reward_signals = {}
     ppo_optimizer.return_value = mock_optimizer
 
-    trainer = PPOTrainer(brain_params, 0, dummy_config, True, False, 0, "0")
+    trainer = PPOTrainer("test_policy", 0, dummy_config, True, False, 0, "0")
     policy = mock.Mock(spec=NNPolicy)
     policy.get_current_step.return_value = 2000
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(brain_params.brain_name)
+
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
     trainer.add_policy(behavior_id, policy)
-    assert trainer.get_policy(brain_params.brain_name) == policy
+    assert trainer.get_policy("test_policy") == policy
 
     # Make sure the summary steps were loaded properly
     assert trainer.get_step == 2000
@@ -359,23 +332,7 @@ def test_add_get_policy(ppo_optimizer, dummy_config):
     # Test incorrect class of policy
     policy = mock.Mock()
     with pytest.raises(RuntimeError):
-        trainer.add_policy(brain_params, policy)
-
-
-# TODO: Move this to test_settings.py
-def test_bad_config():
-    brain_params = make_brain_parameters(
-        discrete_action=False, visual_inputs=0, vec_obs_size=6
-    )
-    # Test that we throw an error if we have sequence length greater than batch size
-    with pytest.raises(TrainerConfigError):
-        TrainerSettings(
-            network_settings=NetworkSettings(
-                memory=NetworkSettings.MemorySettings(sequence_length=64)
-            ),
-            hyperparameters=PPOSettings(batch_size=32),
-        )
-        _ = PPOTrainer(brain_params, 0, dummy_config, True, False, 0, "0")
+        trainer.add_policy(behavior_id, policy)
 
 
 if __name__ == "__main__":