+++ ml-agents/mlagents/trainers/tests/test_sac.py
@@ -10,7 +10,7 @@ from mlagents.trainers.sac.optimizer import SACOptimizer
 from mlagents.trainers.policy.nn_policy import NNPolicy
 from mlagents.trainers.agent_processor import AgentManagerQueue
 from mlagents.trainers.tests import mock_brain as mb
-from mlagents.trainers.tests.mock_brain import make_brain_parameters
+from mlagents.trainers.tests.mock_brain import setup_test_behavior_specs
 from mlagents.trainers.tests.test_trajectory import make_fake_trajectory
 from mlagents.trainers.tests.test_simple_rl import SAC_CONFIG
 from mlagents.trainers.settings import NetworkSettings
@@ -24,7 +24,7 @@ def dummy_config():
     return copy.deepcopy(SAC_CONFIG)
 
 
-VECTOR_ACTION_SPACE = [2]
+VECTOR_ACTION_SPACE = 2
 VECTOR_OBS_SPACE = 8
 DISCRETE_ACTION_SPACE = [3, 3, 3, 2]
 BUFFER_INIT_SAMPLES = 64
@@ -32,12 +32,13 @@ NUM_AGENTS = 12
 
 
 def create_sac_optimizer_mock(dummy_config, use_rnn, use_discrete, use_visual):
-    mock_brain = mb.setup_mock_brain(
+    mock_brain = mb.setup_test_behavior_specs(
         use_discrete,
         use_visual,
-        vector_action_space=VECTOR_ACTION_SPACE,
-        vector_obs_space=VECTOR_OBS_SPACE,
-        discrete_action_space=DISCRETE_ACTION_SPACE,
+        vector_action_space=DISCRETE_ACTION_SPACE
+        if use_discrete
+        else VECTOR_ACTION_SPACE,
+        vector_obs_space=VECTOR_OBS_SPACE if not use_visual else 0,
     )
     trainer_settings = dummy_config
     trainer_settings.network_settings.memory = (
@@ -62,7 +63,9 @@ def test_sac_optimizer_update(dummy_config, rnn, visual, discrete):
         dummy_config, use_rnn=rnn, use_discrete=discrete, use_visual=visual
     )
     # Test update
-    update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, optimizer.policy.behavior_spec
+    )
     # Mock out reward signal eval
     update_buffer["extrinsic_rewards"] = update_buffer["environment_rewards"]
     optimizer.update(
@@ -84,7 +87,9 @@ def test_sac_update_reward_signals(
     )
 
     # Test update, while removing PPO-specific buffer elements.
-    update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, optimizer.policy.brain)
+    update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, optimizer.policy.behavior_spec
+    )
 
     # Mock out reward signal eval
     update_buffer["extrinsic_rewards"] = update_buffer["environment_rewards"]
@@ -95,52 +100,45 @@ def test_sac_update_reward_signals(
 
 
 def test_sac_save_load_buffer(tmpdir, dummy_config):
-    mock_brain = mb.setup_mock_brain(
+    mock_specs = mb.setup_test_behavior_specs(
         False,
         False,
         vector_action_space=VECTOR_ACTION_SPACE,
         vector_obs_space=VECTOR_OBS_SPACE,
-        discrete_action_space=DISCRETE_ACTION_SPACE,
     )
     trainer_params = dummy_config
     trainer_params.hyperparameters.save_replay_buffer = True
-    trainer = SACTrainer(
-        mock_brain.brain_name, 1, trainer_params, True, False, 0, "testdir"
-    )
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(mock_brain.brain_name)
-    policy = trainer.create_policy(behavior_id, mock_brain)
+    trainer = SACTrainer("test", 1, trainer_params, True, False, 0, "testdir")
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
+    policy = trainer.create_policy(behavior_id, mock_specs)
     trainer.add_policy(behavior_id, policy)
 
-    trainer.update_buffer = mb.simulate_rollout(BUFFER_INIT_SAMPLES, policy.brain)
+    trainer.update_buffer = mb.simulate_rollout(
+        BUFFER_INIT_SAMPLES, policy.behavior_spec
+    )
     buffer_len = trainer.update_buffer.num_experiences
     trainer.save_model()
 
     # Wipe Trainer and try to load
-    trainer2 = SACTrainer(
-        mock_brain.brain_name, 1, trainer_params, True, True, 0, "testdir"
-    )
+    trainer2 = SACTrainer("test", 1, trainer_params, True, True, 0, "testdir")
 
-    policy = trainer2.create_policy(behavior_id, mock_brain)
+    policy = trainer2.create_policy(behavior_id, mock_specs)
     trainer2.add_policy(behavior_id, policy)
     assert trainer2.update_buffer.num_experiences == buffer_len
 
 
 @mock.patch("mlagents.trainers.sac.trainer.SACOptimizer")
 def test_add_get_policy(sac_optimizer, dummy_config):
-    brain_params = make_brain_parameters(
-        discrete_action=False, visual_inputs=0, vec_obs_size=6
-    )
     mock_optimizer = mock.Mock()
     mock_optimizer.reward_signals = {}
     sac_optimizer.return_value = mock_optimizer
 
-    trainer = SACTrainer(brain_params, 0, dummy_config, True, False, 0, "0")
+    trainer = SACTrainer("test", 0, dummy_config, True, False, 0, "0")
     policy = mock.Mock(spec=NNPolicy)
     policy.get_current_step.return_value = 2000
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(brain_params.brain_name)
-
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
     trainer.add_policy(behavior_id, policy)
-    assert trainer.get_policy(brain_params.brain_name) == policy
+    assert trainer.get_policy(behavior_id.behavior_id) == policy
 
     # Make sure the summary steps were loaded properly
     assert trainer.get_step == 2000
@@ -152,15 +150,15 @@ def test_add_get_policy(sac_optimizer, dummy_config):
 
 
 def test_advance(dummy_config):
-    brain_params = make_brain_parameters(
-        discrete_action=False, visual_inputs=0, vec_obs_size=6
+    specs = setup_test_behavior_specs(
+        use_discrete=False, use_visual=False, vector_action_space=2
     )
     dummy_config.hyperparameters.steps_per_update = 20
     dummy_config.hyperparameters.reward_signal_steps_per_update = 20
     dummy_config.hyperparameters.buffer_init_steps = 0
-    trainer = SACTrainer(brain_params, 0, dummy_config, True, False, 0, "0")
-    policy = trainer.create_policy(brain_params.brain_name, brain_params)
-    behavior_id = BehaviorIdentifiers.from_name_behavior_id(brain_params.brain_name)
+    trainer = SACTrainer("test", 0, dummy_config, True, False, 0, "0")
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
+    policy = trainer.create_policy(behavior_id, specs)
     trainer.add_policy(behavior_id, policy)
 
     trajectory_queue = AgentManagerQueue("testbrain")
@@ -170,10 +168,9 @@ def test_advance(dummy_config):
 
     trajectory = make_fake_trajectory(
         length=15,
+        observation_shapes=specs.observation_shapes,
         max_step_complete=True,
-        vec_obs_size=6,
-        num_vis_obs=0,
-        action_space=[2],
+        action_space=2,
         is_discrete=False,
     )
     trajectory_queue.put(trajectory)
@@ -190,10 +187,9 @@ def test_advance(dummy_config):
     # Add a terminal trajectory
     trajectory = make_fake_trajectory(
         length=6,
+        observation_shapes=specs.observation_shapes,
         max_step_complete=False,
-        vec_obs_size=6,
-        num_vis_obs=0,
-        action_space=[2],
+        action_space=2,
         is_discrete=False,
     )
     trajectory_queue.put(trajectory)
@@ -216,10 +212,9 @@ def test_advance(dummy_config):
     # two updates, there should NOT be a policy on the queue.
     trajectory = make_fake_trajectory(
         length=5,
+        observation_shapes=specs.observation_shapes,
         max_step_complete=False,
-        vec_obs_size=6,
-        num_vis_obs=0,
-        action_space=[2],
+        action_space=2,
         is_discrete=False,
     )
     trajectory_queue.put(trajectory)
@@ -229,7 +224,8 @@ def test_advance(dummy_config):
 
     # Call add_policy and check that we update the correct number of times.
     # This is to emulate a load from checkpoint.
-    policy = trainer.create_policy(brain_params.brain_name, brain_params)
+    behavior_id = BehaviorIdentifiers.from_name_behavior_id(trainer.brain_name)
+    policy = trainer.create_policy(behavior_id, specs)
     policy.get_current_step = lambda: 200
     trainer.add_policy(behavior_id, policy)
     trainer.optimizer.update = mock.Mock()